+++ src/validation/decorators/is-nullable.decorator.ts
@@ -1,0 +1,16 @@
+import { ValidateIf, ValidationOptions } from 'class-validator';
+
+/**
+ * Skips validation if the property is `null`.
+ */
+export function IsNullable(options?: ValidationOptions): PropertyDecorator {
+  return function IsNullableDecorator(
+    prototype: object,
+    propertyKey: string | symbol,
+  ) {
+    ValidateIf((obj) => obj[propertyKey] !== null, options)(
+      prototype,
+      propertyKey,
+    );
+  };
+}